requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
chromadb
google-genai
pydantic
